Ayo tetap Malas

+++ src/models/mod.rs
@@ -3,7 +3,7 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Serialize, Deserialize, Debug, FromRow)]
 pub struct Users {
-  pub id: i32,
+  pub id: i64,
   pub nama: String,
   pub email: String,
   pub sandi: String,
